Uncomment check auth

+++ controllers/posts.js
@@ -6,7 +6,7 @@ exports.createPost = (req, res, next) => {
     title: req.body.title,
     content: req.body.content,
     imagePath: url + '/images/' + req.file.filename,
-    // creator: req.userData.userId,
+    creator: req.userData.userId,
   });
   post
     .save()
@@ -39,10 +39,10 @@ exports.updatePost = (req, res, next) => {
     title: req.body.title,
     content: req.body.content,
     imagePath: imagePath,
-    // creator: req.userData.userId,
+    creator: req.userData.userId,
   });
-  // Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post)
-  Post.updateOne({ _id: req.params.id }, post)
+  Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post)
+    // Post.updateOne({ _id: req.params.id }, post)
 
     .then((result) => {
       console.log(result);
@@ -103,8 +103,8 @@ exports.getPost = (req, res, next) => {
 };
 
 exports.deletePost = (req, res, next) => {
-  // Post.deleteOne({ _id: req.params.id, creator: req.userData.userId })
-  Post.deleteOne({ _id: req.params.id })
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId })
+    // Post.deleteOne({ _id: req.params.id })
 
     .then((result) => {
       console.log(result);
